+++ ios/RCTBridge+Generics.swift
@@ -1,0 +1,6 @@
+extension RCTBridge {
+    /// Helper to allow acquiring typed module instance from the bridge.
+    subscript<T: RCTBridgeModule>(_ moduleType: T.Type) -> T? {
+        module(for: moduleType) as? T
+    }
+}