+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [0.7.0] (2023-06-26)
+
+### Added
+
+- Support for `SourceConfig.metadata` and `Source.metadata` to allow passing custom metadata to the Player UI
+- Support for `CustomMessageHandler` to to allow two-way communication between the Player UI and the React Native app
+
+### Fixed
+
+- Android module build issues
+
 ## [0.6.0] (2023-03-27)
 
 ### Added